# uni_core/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  app.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
  request.DockerCmdReq:
    type: object
...
  response.DockerCmdResp:
    properties:
      ref:
        description: Ref 是指令号：前端用它轮询结果。
        type: string
    type: object
  response.DockerCmdResultResp:
    properties:
      alreadyExists:
        type: boolean
      detail:
        description: Detail 是排障细节（**页面不渲染**；留给「复制诊断信息」这类入口）。
        type: string
      error:
        description: Error 是终态结论句（页面直接显示）。
        type: string
      payload:
        description: Payload 是结果数据面（按 action 形态不同；前端按 action 解析）。
      sessionId:
        type: string
      status:
        description: Status ∈ pending | succeeded | failed | timeout。
        type: string
      streamTicket:
        description: |-
          StreamTicket 是一次性流票据（三期）：指令建立了流会话时，**每次**轮询响应
          都新签一张（TTL 30s、单次使用，绑定 userId+sessionId+deviceId）。前端接流
          时用**最新**的那张；票据不进日志、不上报，只在该响应里出现。
        type: string
    type: object
  response.DockerComposeInfoResp:
    properties:
      flavor:
        type: string
      version:
        type: string
    type: object
  response.DockerContainerItem:
    properties:
      composeProject:
        type: string
      composeService:
        type: string
      cpuPercent:
        type: number
      created:
        type: integer
      id:
        type: string
      image:
        type: string
      memLimitMb:
        type: number
      memUsageMb:
        type: number
      name:
        type: string
      netRxBytesSec:
        type: number
      netTxBytesSec:
        type: number
      ports:
        items:
          $ref: '#/definitions/response.DockerPortItem'
        type: array
      protected:
        type: boolean
      startedAt:
        type: integer
      state:
        type: string
      statusText:
        type: string
    type: object
  response.DockerFleetContainers:
    properties:
      protected:
        description: Protected 是受保护容器的数量（不论运行态；结论由 agent 按保护清单算好）。
        type: integer
      running:
        description: Running 是 state=running 的容器数。
        type: integer
      stopped:
        description: |-
          Stopped = Total - Running：泛指一切非 running（exited/created/paused/dead…），
          与异常清单的收录口径同一句话（两边各写一套就又出现一次口径分裂）。
        type: integer
      total:
        type: integer
    type: object
  response.DockerFleetDisk:
    properties:
      buildCacheMb:
        type: number
      hosts:
        description: |-
          Hosts 是**有 df 数据**的主机数（快照 disk_usage 非 nil 的主机数）。与
          fleet.hosts.total 并排读就是「4 台里 3 台报了磁盘账」—— 没报的那台不是
          0 占用，是不知道；计数如实标注而不是把缺报主机折算成零。
        type: integer
      imagesDanglingMb:
        description: |-
          ImagesDanglingMB 是悬空镜像占用合计（image:prune 默认目标的体积：
          「能安全收回多少」的镜像半边）。卷的可回收数不在此列 —— volume:prune 只回收
          匿名未用卷，合计口径在协议上拿不到（df 不区分匿名），入口在卷页的清理流。
        type: number
      imagesTotalMb:
        description: ImagesTotalMB / VolumesTotalMB / BuildCacheMB 是三类占用合计（MB）。
        type: number
      volumesTotalMb:
        type: number
    type: object
  response.DockerFleetHosts:
    properties:
      dockerOk:
        description: DockerOK 是 dockerOk=true 的主机数（agent 自报的能力信号）。
        type: integer
      total:
        description: Total 是本响应 hosts 列表的行数（读失败的主机也算一台 —— 行还在，err 如实）。
        type: integer
    type: object
  response.DockerFleetImages:
    properties:
      total:
        type: integer
      unused:
        type: integer
    type: object
  response.DockerFleetNetworks:
    properties:
      total:
        type: integer
    type: object
  response.DockerFleetProjects:
    properties:
      running:
        description: |-
          Running 是 state=running 的项目数（项目态 ∈ running|partial|stopped，
          由 agent 按成员容器归纳 —— 非 running 项目不上异常清单，那是容器粒度的清单）。
        type: integer
      total:
        type: integer
    type: object
  response.DockerFleetVolumes:
    properties:
      total:
        type: integer
      unused:
        type: integer
    type: object
  response.DockerHostDiskItem:
    properties:
      buildCacheMb:
        type: number
      danglingImages:
        description: |-
          DanglingImages 是悬空镜像计数（与镜像页「可回收」计数同一判据：agent 的
          Dangling 结论，前端不重算）。
        type: integer
      imagesDanglingMb:
        description: ImagesDanglingMB 是悬空镜像占用（image:prune 默认目标的体积）。
        type: number
      imagesMb:
        description: ImagesMB / VolumesMB / BuildCacheMB 是三类占用（MB；system df 的同帧口径）。
        type: number
      unusedVolumes:
        description: |-
          UnusedVolumes 是未被任何容器挂载的卷计数。注意它**不是** volume:prune 的
          回收数：prune 只收匿名未用卷，命名卷走卷页的独立删除流 —— 面板只给计数
          与入口，清理动作在列表页的确认档流里发生。
        type: integer
      volumesMb:
        type: number
    type: object
  response.DockerHostItem:
    properties:
      composeFlavor:
        type: string
      composeVersion:
        type: string
      containers:
        type: integer
      disk:
        allOf:
        - $ref: '#/definitions/response.DockerHostDiskItem'
        description: |-
          Disk 是该主机的磁盘占用账目（system df 汇总 + 可回收计数）。**nil = 无 df 数据**
          （那帧 df 失败、或旧版 agent 不发 disk_usage）—— 页面显示「数据不可用」而不是 0
          （与卷条目 SizeMB=nil 同一纪律：不知道 ≠ 零）。只在总览响应填充（磁盘面板的
          专供数据面）；主机清单页不消费它，omitempty 让字段在那边缺席。
      dockerOk:
        description: DockerOK 是 agent 自报的能力信号；false 时前端不渲染任何操作按钮。
        type: boolean
      error:
        description: Error 是 dockerOK=false 的结论句。
        type: string
      hostname:
        type: string
      id:
        example: "0"
        type: string
      images:
        type: integer
      lastSync:
        description: LastSync 是 core 收到快照的时刻（unix 秒）；0 = 从未上报。
        type: integer
      online:
        description: Online 由 last_seen_at 与 sys.agent.offlineThreshold 折算（与设备列表同口径）。
        type: boolean
      primaryIp:
        type: string
      stale:
        description: Stale 是服务端算好的结论（>max(3×周期, 90s) 未更新）。
        type: boolean
    type: object
  response.DockerHostListResp:
    properties:
      list:
        items:
          $ref: '#/definitions/response.DockerHostItem'
        type: array
      snapshotInterval:
        description: |-
          SnapshotInterval 是当前快照周期（秒）：前端据此算「同步于 N 秒前」的文案，
          与后端 stale 判定同源（各自硬编码会让两边对「多久算陈旧」给出不同答案）。
        type: integer
    type: object
  response.DockerImageItem:
    properties:
      created:
        type: integer
      dangling:
        type: boolean
      id:
        type: string
      inUse:
        type: boolean
      inUseBy:
        items:
          type: string
        type: array
      repoTags:
        items:
          type: string
        type: array
      sizeMb:
        type: number
    type: object
  response.DockerNetworkItem:
    properties:
      containersCount:
        type: integer
      driver:
        type: string
      internal:
        type: boolean
      name:
        type: string
      scope:
        type: string
    type: object
  response.DockerOverviewAnomalies:
    properties:
      items:
        description: Items 按主机 id 升序（同机按容器名）排序。
        items:
          $ref: '#/definitions/response.DockerOverviewAnomalyItem'
        type: array
      total:
        description: |-
          Total 是**截断前**的非 running 容器全量数；Items 至多 50 条。
          两者并存：控制塔要一眼知道「舰队里有几个不正常的容器」，详情只给前 50 ——
          超限时 Total 与 Items 长度不等，即「截断了」的显式口径（不靠前端猜）。
        type: integer
    type: object
  response.DockerOverviewAnomalyItem:
    properties:
      hostId:
        example: "0"
        type: string
      hostname:
        type: string
      id:
        description: |-
          ID 是容器的完整 ID（同 DockerContainerItem.ID）：异常行点进去要能深链到
          容器详情页（/docker/container-detail/:id），没有它前端只能干列名字。
        type: string
      image:
        type: string
      name:
        type: string
      protected:
        description: |-
          Protected 是 agent 按保护清单算好的结论：控制塔**不**据此拦截操作（拦在
          指令受理面），只用来让运维一眼看出「这是个动不得的容器」。
        type: boolean
      state:
        type: string
      statusText:
        description: StatusText 是 docker ps 的原生状态句（"Exited (137) 3 days ago"），展示层直接显示。
        type: string
    type: object
  response.DockerOverviewFleet:
    properties:
      containers:
        $ref: '#/definitions/response.DockerFleetContainers'
      disk:
        allOf:
        - $ref: '#/definitions/response.DockerFleetDisk'
        description: |-
          Disk 是磁盘占用的航队汇总（6a 磁盘治理）。值形态而非指针：fleet 恒在
          （零 df 主机时报 Hosts=0 与零合计 —— 「全部主机都没有 df 数据」本身就是一个
          如实的舰队结论，页面上磁贴显示「数据不可用」）。
      hosts:
        $ref: '#/definitions/response.DockerFleetHosts'
      images:
        $ref: '#/definitions/response.DockerFleetImages'
      networks:
        $ref: '#/definitions/response.DockerFleetNetworks'
      projects:
        $ref: '#/definitions/response.DockerFleetProjects'
      volumes:
        $ref: '#/definitions/response.DockerFleetVolumes'
    type: object
  response.DockerOverviewResp:
    properties:
      anomalies:
        allOf:
        - $ref: '#/definitions/response.DockerOverviewAnomalies'
        description: Anomalies 是非 running 容器清单；Total 是截断前的全量数。
      fleet:
        $ref: '#/definitions/response.DockerOverviewFleet'
      hosts:
        items:
          $ref: '#/definitions/response.DockerHostItem'
        type: array
    type: object
  response.DockerPortItem:
    properties:
      ip:
        type: string
      privatePort:
        type: integer
      publicPort:
        type: integer
      type:
        type: string
    type: object
  response.DockerProjectItem:
    properties:
      configFiles:
        items:
          type: string
        type: array
      containersCount:
        type: integer
      name:
        type: string
      protected:
        description: Protected 来自 agent 的结论（project:<名>）；服务粒度由容器条目承载。
        type: boolean
      services:
        type: integer
      state:
        type: string
    type: object
...
  response.DockerStateResp:
    properties:
      ageSeconds:
        description: AgeSeconds 是距上次上报的秒数（0 = 刚刚）。
        type: integer
      compose:
        $ref: '#/definitions/response.DockerComposeInfoResp'
      containers:
        items:
          $ref: '#/definitions/response.DockerContainerItem'
        type: array
      dockerOk:
        type: boolean
      error:
        type: string
      images:
        items:
          $ref: '#/definitions/response.DockerImageItem'
        type: array
      lastSync:
        type: integer
      networks:
        items:
          $ref: '#/definitions/response.DockerNetworkItem'
        type: array
      neverReported:
        description: NeverReported 表示该主机从未上报过快照（与「陈旧」是两句不同的话）。
        type: boolean
      projects:
        items:
          $ref: '#/definitions/response.DockerProjectItem'
        type: array
      stale:
        type: boolean
      volumes:
        items:
          $ref: '#/definitions/response.DockerVolumeItem'
        type: array
    type: object
  response.DockerTaskItem:
    properties:
      action:
        description: |-
          Action 是动作码（前端凭它判断「可流」：image:pull/compose:pull 可开进度流，
          不需要额外的 streamable 字段 —— action 就是判定依据）。
        type: string
      createdAt:
        description: CreatedAt 是受理时刻（unix 毫秒，与指令记录同单位）。
        type: integer
      hostId:
        description: HostID 是目标主机的设备 ID（雪花值，string 编码与全站同纪律）。
        example: "0"
        type: string
      hostname:
        description: Hostname 是目标主机名（联查设备表；设备已删时为空 —— hostId 仍是可导航主键）。
        type: string
      ref:
        description: Ref 是指令号（轮询 /cmds/:ref 与拉取进度流 /cmds/:ref/pull 的钥匙）。
        type: string
      status:
        description: Status 终态照实投影：pending 在执行、succeeded 成功、failed 失败、timeout 超时。
        type: string
      summary:
        description: |-
          Summary 是终态结论句原文（pending 时为空；成功兜底「执行成功」——与审计
          入账的结论句同一句话，见 service/docker_audit.go 的 terminalSummary）。
        type: string
      target:
        description: Target 是操作目标（镜像名/容器名/项目名…）。
        type: string
      username:
        description: Username 是发起人用户名（联查用户表；用户已删时为空，审计仍按 UserID 记账）。
        type: string
    type: object
  response.DockerTaskListResp:
    properties:
      items:
        items:
          $ref: '#/definitions/response.DockerTaskItem'
        type: array
    type: object
  response.DockerVolumeItem:
    properties:
      driver:
        type: string
      inUse:
        type: boolean
      mountedBy:
        items:
          type: string
        type: array
      name:
        type: string
      protected:
        description: |-
          Protected 来自 agent 的结论（volume:<名>），前端不重复实现判断。
          不带 omitempty：与 DockerContainerItem.Protected 同一形态（0 是真实值）。
        type: boolean
      sizeMb:
        type: number
    type: object
  response.DockerWorkloadItem:
    properties:
      composeProject:
        type: string
      composeService:
        type: string
      cpuPercent:
        type: number
      created:
        type: integer
      hostId:
        description: HostID 是容器所在主机的设备 ID：统一表的核心新增（容器在哪台机的归属）。
        example: "0"
        type: string
      hostname:
        description: Hostname 是容器所在主机的主机名（与主机清单/总览的 hostname 同源）。
        type: string
      id:
        type: string
      image:
        type: string
      memLimitMb:
        type: number
      memUsageMb:
        type: number
      name:
        type: string
      netRxBytesSec:
        type: number
      netTxBytesSec:
        type: number
      ports:
        items:
          $ref: '#/definitions/response.DockerPortItem'
        type: array
      protected:
        type: boolean
      startedAt:
        type: integer
      state:
        type: string
      statusText:
        type: string
    type: object
  response.DockerWorkloadListResp:
    properties:
      items:
        items:
          $ref: '#/definitions/response.DockerWorkloadItem'
        type: array
      total:
        description: |-
          Total 是**截断前**的全量数（过滤后）；Items 至多 500 条。两者并存的口径
          与总览 anomalies 同一句话：Total 与 Items 长度不等即「截断了」的显式信号
          （不靠前端猜「还有多少」）。
        type: integer
    type: object
...
info:
  contact:
    email: support@example.com
    name: API Support
  description: 企业级 RBAC 权限管理系统后端 API
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: RBAC Admin API
  version: 1.0.0
paths:
...
  /docker/containers:
    get:
      description: 全部可管主机的容器并成一张表（每行带 hostId/hostname 归属）；hostId 限定单主机、state 过滤运行态（stopped=一切非
        running，与总览 KPI 同口径）、keyword 按容器名或镜像名子串匹配；上限 500 条、total 如实报截断前全量；单台快照读失败跳过该主机（其故障在总览页如实呈现）
      parameters:
      - description: 限定单主机(缺省=全部可管主机)
        format: int64
        in: query
        name: hostId
        type: integer
      - description: 运行态过滤(running/stopped)
        in: query
        name: state
        type: string
      - description: 容器名或镜像名子串(大小写不敏感)
        in: query
        name: keyword
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 查询成功
          schema:
            allOf:
            - $ref: '#/definitions/app.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.DockerWorkloadListResp'
              type: object
        "400":
          description: 参数错误(state 非 running/stopped)
          schema:
            $ref: '#/definitions/app.Response'
        "401":
          description: 未登录
          schema:
            $ref: '#/definitions/app.Response'
        "403":
          description: 无权限
          schema:
            $ref: '#/definitions/app.Response'
      security:
      - BearerAuth: []
      summary: 跨主机容器统一表
      tags:
      - Docker 管理
...
  /docker/hosts:
    get:
      description: 枚举上报过快照的主机（docker:hosts 集合）及各自的摘要与陈旧结论
      produces:
      - application/json
      responses:
        "200":
          description: 查询成功
          schema:
            allOf:
            - $ref: '#/definitions/app.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.DockerHostListResp'
              type: object
        "401":
          description: 未登录
          schema:
            $ref: '#/definitions/app.Response'
        "403":
          description: 无权限
          schema:
            $ref: '#/definitions/app.Response'
      security:
      - BearerAuth: []
      summary: 可管主机清单
      tags:
      - Docker 管理
  /docker/hosts/{id}/cmds:
    post:
      consumes:
      - application/json
      description: 校验 action 对应的权限码与期次闸后下发；立即返回指令号（ref），结果靠轮询
      parameters:
      - description: 设备ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      - description: 指令请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.DockerCmdReq'
      produces:
      - application/json
      responses:
        "202":
          description: 已受理（data.ref 为指令号）
          schema:
            allOf:
            - $ref: '#/definitions/app.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.DockerCmdResp'
              type: object
        "400":
          description: 参数错误 / 未知操作 / 该操作尚未开放
          schema:
            $ref: '#/definitions/app.Response'
        "401":
          description: 未登录
          schema:
            $ref: '#/definitions/app.Response'
        "403":
          description: 无操作权限（force=true 时还要求 docker:exec 级）
          schema:
            $ref: '#/definitions/app.Response'
        "409":
          description: 该目标上已有同一条指令在执行
          schema:
            $ref: '#/definitions/app.Response'
        "500":
          description: 设备离线 / docker 不可用 / 内部错误
          schema:
            $ref: '#/definitions/app.Response'
      security:
      - BearerAuth: []
      summary: 受理 docker 指令
      tags:
      - Docker 管理
  /docker/hosts/{id}/cmds/{ref}:
    get:
      description: 按指令号查询状态与载荷（pending/running/succeeded/failed/timeout）；建立了流会话时附带一次性
        streamTicket
      parameters:
      - description: 设备ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      - description: 指令号
        in: path
        name: ref
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 查询成功
          schema:
            allOf:
            - $ref: '#/definitions/app.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.DockerCmdResultResp'
              type: object
        "401":
          description: 未登录
          schema:
            $ref: '#/definitions/app.Response'
        "403":
          description: 无操作权限 / 非发起人
          schema:
            $ref: '#/definitions/app.Response'
        "404":
          description: 指令不存在或已过期
          schema:
            $ref: '#/definitions/app.Response'
      security:
      - BearerAuth: []
      summary: 轮询 docker 指令结果
      tags:
      - Docker 管理
...
  /docker/hosts/{id}/state:
    get:
      description: 返回一台主机的容器/镜像/卷/网络/编排项目清单与陈旧结论（离线可读，只标 stale）
      parameters:
      - description: 设备ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 查询成功
          schema:
            allOf:
            - $ref: '#/definitions/app.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.DockerStateResp'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/app.Response'
        "401":
          description: 未登录
          schema:
            $ref: '#/definitions/app.Response'
        "403":
          description: 无权限
          schema:
            $ref: '#/definitions/app.Response'
        "404":
          description: 设备不存在
          schema:
            $ref: '#/definitions/app.Response'
      security:
      - BearerAuth: []
      summary: 主机资源快照
      tags:
      - Docker 管理
...
  /docker/overview:
    get:
      description: 聚合全部可管主机的容器/镜像/卷/网络/项目计数与异常容器清单（条目与 docker/hosts 同形态；单台快照读失败只降级该主机并如实标注
        error）
      produces:
      - application/json
      responses:
        "200":
          description: 查询成功
          schema:
            allOf:
            - $ref: '#/definitions/app.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.DockerOverviewResp'
              type: object
        "401":
          description: 未登录
          schema:
            $ref: '#/definitions/app.Response'
        "403":
          description: 无权限
          schema:
            $ref: '#/definitions/app.Response'
      security:
      - BearerAuth: []
      summary: 控制塔总览
      tags:
      - Docker 管理
...
  /docker/tasks:
    get:
      description: 最近受理的 docker 指令（≤100 条、受理时刻降序、跨主机聚合）；hostId 限定单主机、status 过滤 pending/done、action
        过滤动作码；条目含 ref/主机/动作/目标/发起人用户名/受理时刻/终态（pending/succeeded/failed/timeout）/终态结论句；拉取类任务前端凭
        action 复用 /cmds/:ref/pull 打开进度流（取消=断开进度流）
      parameters:
      - description: 限定单主机(缺省=跨主机)
        format: int64
        in: query
        name: hostId
        type: integer
      - description: 阶段过滤(pending/done)
        in: query
        name: status
        type: string
      - description: 动作码过滤(如 image:pull)
        in: query
        name: action
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 查询成功
          schema:
            allOf:
            - $ref: '#/definitions/app.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.DockerTaskListResp'
              type: object
        "400":
          description: 参数错误(status 非 pending/done) / 未知操作
          schema:
            $ref: '#/definitions/app.Response'
        "401":
          description: 未登录
          schema:
            $ref: '#/definitions/app.Response'
        "403":
          description: 无权限(docker:list)
          schema:
            $ref: '#/definitions/app.Response'
      security:
      - BearerAuth: []
      summary: 最近任务
      tags:
      - Docker 管理
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"